+++ Neural_network/neuron.h
@@ -8,15 +8,15 @@
 
 typedef struct neuron_t
 {
-	float actv;
-	float *out_weights;
-	float bias;
-	float z;
+	double actv;
+	double *out_weights;
+	double bias;
+	double z;
 
-	float dactv;
-	float *dw;
-	float dbias;
-	float dz;
+	double dactv;
+	double *dw;
+	double dbias;
+	double dz;
 
 } neuron;
